modularize user popover
