+++ frontend/components/NavbarUserPopover/NavbarUserPopover.module.css
@@ -1,0 +1,39 @@
+.adminBadge {
+    margin-right: var(--spacingHorizontalXS);
+    margin-left: var(--spacingHorizontalS);
+    position: relative;
+    bottom: var(--spacingVerticalXXS);
+}
+
+.popoverButtonContainer {
+    display: flex;
+    width: 100%;
+    flex-direction: row;
+    justify-content: space-between;
+    gap: var(--spacingVerticalS);
+    align-items: center;
+    margin-bottom: var(--spacingVerticalS);
+}
+
+.popoverLink {
+    display: block;
+    margin-bottom: var(--spacingVerticalS);
+}
+
+.popoverButton {
+    margin: 0 !important;
+    padding: 0 !important;
+    min-width: 0 !important;
+}
+
+.popoverLeft {
+    position: relative;
+    bottom: var(--spacingVerticalM);
+    left: var(--spacingHorizontalS);
+}
+
+.popoverRight {
+    position: relative;
+    bottom: var(--spacingVerticalM);
+    right: var(--spacingHorizontalS);
+}
